requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
anytree==2.9.0
    # via pipgrip
beautifulsoup4==4.12.2
    # via -r requirements.in
cachecontrol[filecache]==0.13.1
    # via pip-audit
certifi==2023.7.22
    # via requests
charset-normalizer==3.2.0
    # via requests
click==8.1.4
    # via pipgrip
cyclonedx-python-lib==4.0.1
    # via pip-audit
defusedxml==0.7.1
    # via py-serializable
filelock==3.12.2
    # via cachecontrol
html5lib==1.1
    # via pip-audit
idna==3.4
    # via requests
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
msgpack==1.0.5
    # via cachecontrol
packageurl-python==0.11.1
    # via cyclonedx-python-lib
packaging==23.1
    # via
    #   pip-audit
    #   pip-requirements-parser
    #   pipgrip
pip-api==0.0.30
    # via pip-audit
pip-audit==2.6.0
    # via -r requirements.in
pip-requirements-parser==32.0.1
    # via pip-audit
pipfile==0.0.2
    # via -r requirements.in
pipgrip==0.10.4
    # via -r requirements.in
pkginfo==1.7.1
    # via pipgrip
platformdirs==3.8.1
    # via -r requirements.in
py-serializable==0.11.1
    # via cyclonedx-python-lib
pygments==2.15.1
    # via rich
pyparsing==3.1.0
    # via pip-requirements-parser
python-dateutil==2.8.2 ; python_version < "3.11"
    # via -r requirements.in
requests==2.31.0
    # via
    #   -r requirements.in
    #   cachecontrol
    #   pip-audit
rich==13.4.2
    # via
    #   -r requirements.in
    #   pip-audit
six==1.16.0
    # via
    #   anytree
    #   html5lib
    #   python-dateutil
sortedcontainers==2.4.0
    # via cyclonedx-python-lib
soupsieve==2.4.1
    # via beautifulsoup4
toml==0.10.2
    # via
    #   pip-audit
    #   pipfile
tomlkit==0.12.1 ; python_version < "3.11"
    # via -r requirements.in
urllib3==2.0.3
    # via requests
webencodings==0.5.1
    # via html5lib
wheel==0.40.0
    # via pipgrip

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
